fix: исправление замечаний ревьювера

+++ src/main/java/ru/practicum/shareit/item/service/ItemServiceImpl.java
@@ -9,6 +9,7 @@ import ru.practicum.shareit.item.dto.ItemDto;
 import ru.practicum.shareit.item.model.Item;
 import ru.practicum.shareit.item.storage.ItemStorage;
 import ru.practicum.shareit.user.User;
+import ru.practicum.shareit.user.storage.UserStorage;
 
 import java.util.Collection;
 import java.util.List;
@@ -16,20 +17,35 @@ import java.util.List;
 @Service
 public class ItemServiceImpl implements ItemService {
     private final ItemStorage itemStorage;
+    private final UserStorage userStorage;
 
-    public ItemServiceImpl(ItemStorage itemStorage) {
+    public ItemServiceImpl(ItemStorage itemStorage, UserStorage userStorage) {
         this.itemStorage = itemStorage;
+        this.userStorage = userStorage;
     }
 
     @Override
-    public Item addItem(Item item) {
-        return itemStorage.addNewItem(item)
+    public ItemDto addItem(ItemDto itemDto, Long ownerId)
+            throws AccessDeniedException, InternalServerException {
+        User owner = userStorage.getUserById(ownerId)
+                .orElseThrow(() -> new NotFoundException("Пользователь не найден id=" + ownerId));
+
+        Item newItem = ItemMapper.toItem(itemDto);
+        newItem.setOwner(owner);
+        Item savedItem = itemStorage.addNewItem(newItem)
                 .orElseThrow(() ->
-                        new InternalServerException("Ошибка при добавления вещи : " + item.toString()));
+                        new InternalServerException("Ошибка при добавления вещи : " + newItem));
+        return ItemMapper.toItemDto(savedItem);
     }
 
     @Override
-    public Item updateItem(Item updItem) {
+    public ItemDto updateItem(ItemDto updItemDto, Long ownerId) {
+        User owner = userStorage.getUserById(ownerId)
+                .orElseThrow(() -> new NotFoundException("Пользователь не найден id=" + ownerId));
+
+        Item updItem = ItemMapper.toItem(updItemDto);
+        updItem.setOwner(owner);
+
         Item item = itemStorage.getItemById(updItem.getId())
                 .orElseThrow(() ->
                         new NotFoundException("Не найдена вещь id=" + updItem.getId()));
@@ -50,34 +66,44 @@ public class ItemServiceImpl implements ItemService {
             item.setAvailable(updItem.getAvailable());
         }
 
-        return itemStorage.updateItem(item)
+        Item savedItem = itemStorage.updateItem(item)
                 .orElseThrow(() ->
                         new InternalServerException("Ошибка при обновлении информации: " + updItem));
+        return ItemMapper.toItemDto(savedItem);
     }
 
     @Override
-    public Item getItem(Long id) {
-        return itemStorage.getItemById(id)
+    public ItemDto getItem(Long id) {
+        Item item = itemStorage.getItemById(id)
                 .orElseThrow(() ->
                         new NotFoundException("Не найдена вещь id=" + id));
+        return ItemMapper.toItemDto(item);
     }
 
     @Override
-    public void deleteItem(Long id) {
+    public void deleteItem(Long id, Long ownerId) {
+        Item item = itemStorage.getItemById(id)
+                .orElseThrow(() ->
+                        new NotFoundException("Не найдена вещь id=" + id));
+        if (!item.getOwner().getId().equals(ownerId)) {
+            throw new AccessDeniedException("Удалять вещь может только хозяин.");
+        }
         itemStorage.deleteItem(id);
     }
 
     @Override
-    public String deleteAllItems() {
+    public Collection<ItemDto> deleteAllItems() {
         itemStorage.deleteAllItems();
-        return "Все вещи удалены.";
+        return List.of();
     }
 
     @Override
-    public Collection<ItemDto> getItemsByOwnerId(User owner) {
+    public Collection<ItemDto> getItemsByOwnerId(Long ownerId) {
+        User owner = userStorage.getUserById(ownerId)
+                .orElseThrow(() ->
+                        new NotFoundException("Владелец не найден. id=" + ownerId));
         Collection<ItemDto> items;
-        items = itemStorage.getAllItems().stream()
-                .filter(item -> item.getOwner().equals(owner))
+        items = itemStorage.getAllItemsByOwner(owner).stream()
                 .map(ItemMapper::toItemDto)
                 .toList();
         return items;
@@ -85,18 +111,8 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public Collection<ItemDto> searchItemsByText(final String text) {
-        if (text == null || text.isEmpty()) {
-            return List.of();
-        }
-
-        Collection<ItemDto> items;
-        items = itemStorage.getAllItems().stream()
-                .filter(item ->
-                        item.getName().toUpperCase().contains(text.toUpperCase())
-                                || item.getDescription().toUpperCase().contains(text.toUpperCase()))
-                .filter(Item::getAvailable)
+        return itemStorage.findItemsBytext(text).stream()
                 .map(ItemMapper::toItemDto)
                 .toList();
-        return items;
     }
 }
